feat(profiling): store string data in an arena allocator
This uses `bumpalo::Bump` as an arena allocator to store the string
data contiguously with fewer calls to the system allocator.

Since the `StringTable` owns the arena and stores references to that
data inside other members of the `StringTable`, it is a self-
referencing data structure, which is generally unsafe. This uses
`ouroboros::self_referencing` to have a safe abstraction for this
pattern.